move slate node resetting into effect
