+++ packages/mock-block-dock/package.json
@@ -39,7 +39,7 @@
   "dependencies": {
     "@blockprotocol/core": "0.0.12",
     "@blockprotocol/graph": "0.0.18",
-    "@blockprotocol/hook": "^0.0.5",
+    "@blockprotocol/hook": "0.0.6",
     "@emotion/react": "^11.10.0",
     "@emotion/styled": "^11.10.0",
     "@lit-labs/react": "^1.0.4",
